video 107

+++ src/main/java/guru/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -57,9 +57,14 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID beerId) {
+    public Boolean deleteById(UUID beerId) {
 
-        beerRepository.deleteById(beerId);
+        if(beerRepository.existsById(beerId)){
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+
+        return false;
     }
 
     @Override
